+++ crates/api/src/auth/tests.rs
@@ -1,0 +1,10 @@
+use super::constant_time_eq;
+
+#[test]
+fn compares_equal_and_unequal_inputs() {
+    assert!(constant_time_eq(b"abc", b"abc"));
+    assert!(!constant_time_eq(b"abc", b"abd"));
+    assert!(!constant_time_eq(b"abc", b"ab"));
+    assert!(!constant_time_eq(b"", b"a"));
+    assert!(constant_time_eq(b"", b""));
+}
